refactor(python): Move `read_schema` to top-level `io` module
